# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "File System Service",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "FileSystemItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dir-123"
          },
          "name": {
            "type": "string",
            "example": "MyFolder"
          },
          "parentId": {
            "type": "string",
            "example": "root"
          },
          "type": {
            "type": "string",
            "enum": ["folder", "file"],
            "example": "folder"
          },
          "createdAt": {
            "type": "integer",
            "example": 1767060000
          },
          "updatedAt": {
            "type": "integer",
            "example": 1767060500
          }
        }
      },
      "Folder": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/FileSystemItem"
          },
          {
            "properties": {
              "type": {
                "type": "string",
                "enum": ["folder"]
              }
            }
          }
        ]
      },
      "FileItem": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/FileSystemItem"
          },
          {
            "properties": {
              "type": {
                "type": "string",
                "enum": ["file"]
              },
              "size": {
                "type": "integer",
                "example": 1024
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
    "/filesystem": {
      "get": {
        "summary": "Get root folders (C:, D:)",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "responses": {
          "200": {
            "description": "List of root folders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Folder"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/{parentId}": {
      "get": {
        "summary": "Get contents of a specific folder",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "parentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the parent folder"
          },
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "responses": {
          "200": {
            "description": "Contents of the folder",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileSystemItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/overview": {
      "get": {
        "summary": "Get entire file system for a user",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "responses": {
          "200": {
            "description": "Full file system tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileSystemItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/folder": {
      "post": {
        "summary": "Create a new folder",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["parentId", "name"],
                "properties": {
                  "parentId": {
                    "type": "string",
                    "example": "root"
                  },
                  "name": {
                    "type": "string",
                    "example": "New Folder"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Folder"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/file": {
      "post": {
        "summary": "Create a new file",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["parentId", "name", "size"],
                "properties": {
                  "parentId": {
                    "type": "string",
                    "example": "folder-123"
                  },
                  "name": {
                    "type": "string",
                    "example": "document.txt"
                  },
                  "size": {
                    "type": "integer",
                    "example": 1024
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/{id}/rename": {
      "put": {
        "summary": "Rename a file or folder",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the file or folder to rename"
          },
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newName"],
                "properties": {
                  "newName": {
                    "type": "string",
                    "example": "Renamed Item"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item renamed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileSystemItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing newName)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/{id}": {
      "delete": {
        "summary": "Delete a file or folder",
        "tags": ["File System"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the file or folder to delete"
          },
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID for authorization and ownership"
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted successfully"
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}